Fix mistakes and increase clarity of norm documentation
